Only include the parts of vanilla that we need
This mean the theme will be better encapsulated. All patterns now pull in their own dependencies from vanilla, so if you want to include just one of the pattern files it should work (apart from the settings it depends on - I guess we should sort that in the future sometime).

Also, we aren't including parts of vanilla that we don't need, which should reduce the size of the built CSS, and mean that we reduce the chance of unexpected conflicts.
If the user of the theme wants things that are in vanilla but not needed for this theme, they should import it directly from vanilla.